refactor utils file
