+++ backend/.gitignore
@@ -1,2 +1,3 @@
 dist/*.js
 src/**/*.js
+coverage
